Initial commit for geographical coordinate transformations
Added LLA to ECEF. Which I hope! is right. I used some online calculators to determine if I was, and they gave me the same answer.

Also build an ellipsoid implementation. Only support WGS84 for now, but those are just constants that are defined at the top of the file (semi major axis and reciprocal of flattening). Can add more surveys to allow for different ellipsoid models to be used.