Changed lla2ecef and ecef2lla unit tests to use degrees
To show a more accurate use case, I used lla degrees and used the built in std to_radians() functions for f64's. I want to stay consistent with the types that I use and I'm going to ALWAYS use/return radians. So the user should expect to input radians into any of the helper functions. This test confirms that the math is right even when the user has degrees in lla and converts them to radians.