Support for LLA2NED and LLA2ENU
The process is to convert the origin of the local tangent plane to ECEF and the actual point to ECEF (cartesian) and apply the corresponding rotation matrix (NED or ENU respectively) to (actual - origin). We can confirm this works by providing the same input LLA values for origin of the LTP and actual point to both ENU and NED conversions - since ENU and NED have a simple conversion, it's a simple check to make sure the math is right.